no longer randomizing rand_state
